Remove collapsible text view abstraction
