+++ iOS/Charter/Charter/Thread Detail View Controller/FullEmailMessageTableViewCell.swift
@@ -14,13 +14,13 @@ protocol FullEmailMessageTableViewCellDelegate: class {
     func presentPopover(view: UIView, sender: UIView)
 }
 
-class FullEmailMessageTableViewCell: UITableViewCell, CollapsibleTextViewDataSourceDelegate, RegionViewDelegate {
+class FullEmailMessageTableViewCell: UITableViewCell, EmailTextRegionViewDataSourceDelegate, RegionViewDelegate {
     @IBOutlet weak var nameLabel: UILabel!
     @IBOutlet weak var dateLabel: UILabel!
     @IBOutlet weak var leadingMarginConstraint: NSLayoutConstraint!
     @IBOutlet weak var regionView: RegionView!
     
-    var textViewDataSource: CollapsibleTextViewDataSource? {
+    var textViewDataSource: EmailTextRegionViewDataSource? {
         didSet {
             textViewDataSource?.delegate = self
             regionView.dataSource = textViewDataSource
@@ -51,16 +51,11 @@ class FullEmailMessageTableViewCell: UITableViewCell, CollapsibleTextViewDataSou
         leadingMarginConstraint.constant = CGFloat(indentationLevel) * indentationWidth
     }
     
-    func collapsibleTextViewDataSource(dataSource: CollapsibleTextViewDataSource, didChangeRegionAtIndex index: Int) {
-        let newView = dataSource.regionView(regionView, viewForRegionAtIndex: index)
-        regionView.replaceRegionAtIndex(index, withView: newView)
-    }
-    
     func regionView(regionView: RegionView, didFinishReplacingRegionAtIndex: Int) {
         delegate?.didChangeCellHeight(NSIndexPath(forRow: didFinishReplacingRegionAtIndex, inSection: 0))
     }
     
-    func collapsibleTextViewDataSourceNeedsPopoverViewControllerPresented(view: UIView, sender: UIView) {
+    func emailTextRegionViewDatatSourceNeedsPopoverViewControllerPresented(view: UIView, sender: UIView) {
         delegate?.presentPopover(view, sender: sender)
     }
