+++ iOS/Charter/Charter/Thread Detail View Controller/ThreadDetailDataSource.swift
@@ -26,13 +26,13 @@ class ThreadDetailDataSourceImpl: NSObject, ThreadDetailDataSource {
     
     private var indentationAndEmail: [(Int, Email)] = [] {
         didSet {
-            textViewDataSources = [NSIndexPath: EmailCollapsibleTextViewDataSource]()
+            textViewDataSources = [NSIndexPath: EmailTextRegionViewDataSource]()
         }
     }
     
     private let codeBlockParser: CodeBlockParser
     private let rootEmail: Email
-    private var textViewDataSources: [NSIndexPath: EmailCollapsibleTextViewDataSource] = [NSIndexPath: EmailCollapsibleTextViewDataSource]()
+    private var textViewDataSources: [NSIndexPath: EmailTextRegionViewDataSource] = [NSIndexPath: EmailTextRegionViewDataSource]()
     private lazy var emailFormatter: EmailFormatter = EmailFormatter()
     
     private var emails: [Email] = [] {
@@ -51,13 +51,13 @@ class ThreadDetailDataSourceImpl: NSObject, ThreadDetailDataSource {
     func registerTableView(tableView: UITableView) {
         tableView.registerNib(FullEmailMessageTableViewCell.nib(), forCellReuseIdentifier: cellIdentifier)
         
-        service.getCachedThreads(descendantsRequestForRootEmail(rootEmail)) { [unowned self] (emails) -> Void in
+        service.getCachedThreads(descendantsRequestForRootEmail(rootEmail)) { (emails) -> Void in
             self.emails = emails
             tableView.reloadData()
             
             if self.rootEmail.descendants.count > emails.count {
                 // Get uncached threads if we are missing any
-                self.service.getUncachedThreads(self.descendantsRequestForRootEmail(self.rootEmail), completion: { [unowned self] (descendants) -> Void in
+                self.service.getUncachedThreads(self.descendantsRequestForRootEmail(self.rootEmail), completion: { (descendants) -> Void in
                     self.emails = descendants
                     tableView.reloadData()
                 })
@@ -84,8 +84,8 @@ class ThreadDetailDataSourceImpl: NSObject, ThreadDetailDataSource {
         let content = emailFormatter.formatContent(email.content)
         
         if textViewDataSource == nil {
-            let regions = EmailCollapsibleTextViewDataSource.QuoteRanges(content)
-            textViewDataSource = EmailCollapsibleTextViewDataSource(text: content, initiallyCollapsedRegions: regions, codeBlockParser: codeBlockParser)
+            let regions = EmailQuoteRanges(content)
+            textViewDataSource = EmailTextRegionViewDataSource(text: content, initiallyCollapsedRegions: regions, codeBlockParser: codeBlockParser)
             textViewDataSources[indexPath] = textViewDataSource!
         }
