+++ iOS/Charter/Podfile
@@ -4,9 +4,6 @@ platform :ios, '9.0'
 use_frameworks!
 
 target 'Charter' do
-  pod 'CollapsibleTextView', :path => '../CollapsibleTextView'
-  pod 'GZIP', '~> 1.1.1'
-  pod 'MMMarkdown', '~> 0.5.1'
   pod 'RealmSwift', '~> 0.98.0'
 end
